USVeteranToken  USVT (#1772)
* Add files via upload

* USVeteranToken USVT

USVeteranToken (USVT) is a faster token implementation on a real decentralized immutable blockchain (Ethereum Classic) with usable smart contracts, and is also the first exclusive token on the blockchain for Veterans running on the Ethereum Classic Blockchain.

It is designed to be used as a decentralized ‘Ethereum Classic-like’ token within the Ethereum Classic ecosystem and beyond. It avoids problems related to centralization as seen on Ethereum and security because it is powered by the Ethereum Classic Network and by globally distributed anonymous miners.